+++ jpress-model/src/main/java/io/jpress/model/Attachment.java
@@ -17,6 +17,7 @@ package io.jpress.model;
 
 import com.google.common.collect.Lists;
 import io.jboot.db.annotation.Table;
+import io.jboot.utils.FileUtils;
 import io.jpress.model.base.BaseAttachment;
 
 import java.util.List;
@@ -94,6 +95,20 @@ public class Attachment extends BaseAttachment<Attachment> {
 
     public Attachment(String domainOrPath) {
 
+        setPath(domainOrPath);
+        setSuffix(FileUtils.getSuffix(domainOrPath));
+    }
+
+    /**
+     * 是否是本地附件
+     *
+     * @return
+     */
+    public boolean isLocal() {
+        String path = getPath();
+        return path != null
+                && path.toLowerCase().startsWith("http://") == false
+                && path.toLowerCase().startsWith("https://") == false;
     }
 
 }
